+++ src/math/he.ts
@@ -9,7 +9,7 @@ export default function he(shape: MatrixShape): Matrix {
   const n_in = shape[1];
   const n = shape[0] * n_in;
   const limit = Math.sqrt(6 / n_in);
-  const data = new Float64Array(n);
+  const data = new Float32Array(n);
   
   for (let i = 0; i < n; i++) {
     data[i] = (Math.random() * 2 - 1) * limit;
