+++ src/math/map.ts
@@ -7,7 +7,7 @@ export default function map(
   a: Matrix,
   func: (value: number) => number
 ): Matrix {
-  const result = new Float64Array(a._data.length);
+  const result = new Float32Array(a._data.length);
   for (let i = 0; i < a._data.length; i++) {
     result[i] = func(a._data[i]);
   }
