+++ src/math/ones.ts
@@ -8,7 +8,7 @@ import Matrix from "../matrix";
  */
 export default function ones(shape: MatrixShape): Matrix {
   const n = shape[0] * shape[1];
-  const data = new Float64Array(n);
+  const data = new Float32Array(n);
   // Array.fill sangat dioptimasi di V8
   data.fill(1);
   return Matrix.fromFlat(data, shape);
