+++ src/math/random.ts
@@ -6,7 +6,7 @@ import Matrix from "../matrix";
  */
 export default function random(shape: MatrixShape): Matrix {
   const n = shape[0] * shape[1];
-  const data = new Float64Array(n);
+  const data = new Float32Array(n);
   // Gunakan scale 0.1 agar tidak meledak di model dalam (Xavier-like approximation)
   const scale = 0.1; 
   for (let i = 0; i < n; i++) {
