+++ src/math/sumAxis.ts
@@ -10,7 +10,7 @@ import { isNativeAvailable, sumAxisNative } from "./rust_backend";
 export default function sumAxis(a: Matrix, axis: number, out?: Matrix): Matrix {
   const [rows, cols] = a._shape;
   const outShape: [number, number] = axis === 1 ? [rows, 1] : [1, cols];
-  const result = out || Matrix.fromFlat(new Float64Array(outShape[0] * outShape[1]), outShape);
+  const result = out || Matrix.fromFlat(new Float32Array(outShape[0] * outShape[1]), outShape);
 
   if (isNativeAvailable()) {
     sumAxisNative(a._data, rows, cols, axis, result._data);
